python-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
supabase==2.3.0
python-multipart==0.0.6
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
pytz==2023.3 